Clean up the debug message garbage.
