+++ ENIGMAsystem/SHELL/Widget_Systems/Cocoa/Makefile
@@ -1,3 +1,3 @@
-SOURCES += Widget_Systems/Cocoa/dialogs.cpp
+SOURCES += Widget_Systems/Cocoa/dialogs.cpp Widget_Systems/General/WSdialogs.cpp
 SOURCES += Widget_Systems/Cocoa/dialogs.m
 LDLIBS += -lz -framework Cocoa
