+++ ENIGMAsystem/SHELL/Widget_Systems/General/WSdialogs.cpp
@@ -1,0 +1,24 @@
+#include "Widget_Systems/widgets_mandatory.h"
+
+using std::string;
+
+void show_debug_message_helper(string errortext, MESSAGE_TYPE type);
+
+namespace enigma_user {
+
+void show_debug_message(string errortext, MESSAGE_TYPE type) {
+  if (type != M_INFO && type != M_WARNING) {
+    show_debug_message_helper(errortext, type);
+  } else {
+    #ifndef DEBUG_MODE
+    errortext += "\n";
+    fputs(errortext.c_str(), stderr);
+    fflush(stderr);
+    #endif
+    if (type == MESSAGE_TYPE::M_FATAL_ERROR || 
+      type == MESSAGE_TYPE::M_FATAL_USER_ERROR)
+      abort();
+  }
+}
+
+}
