+++ ENIGMAsystem/SHELL/Widget_Systems/None/default_log.cpp
@@ -20,6 +20,7 @@ if (type == MESSAGE_TYPE::M_USER_ERROR || type == MESSAGE_TYPE::M_FATAL_USER_ERR
   } 
   
   else printf((enigma::error_type(type) + ": %s\n").c_str(), err.c_str());
+  fflush(stdout);
   
   if (type == MESSAGE_TYPE::M_FATAL_ERROR || type == MESSAGE_TYPE::M_FATAL_USER_ERROR) exit(133);
   ABORT_ON_ALL_ERRORS();
