+++ ENIGMAsystem/SHELL/Widget_Systems/Win32/Makefile
@@ -1,3 +1,3 @@
-SOURCES += $(wildcard Widget_Systems/Win32/*.cpp)
+SOURCES += $(wildcard Widget_Systems/Win32/*.cpp) Widget_Systems/General/WSdialogs.cpp
 RESOURCES += $(wildcard Widget_Systems/Win32/*.rc)
 override LDLIBS += -lcomctl32 -luuid -lole32
