+++ ENIGMAsystem/SHELL/Widget_Systems/xlib/Makefile
@@ -1,6 +1,4 @@
-SOURCES += Widget_Systems/xlib/zenity.cpp
-SOURCES += Widget_Systems/xlib/kdialog.cpp
-SOURCES += Widget_Systems/xlib/dialogs.cpp
+SOURCES += $(wildcard Widget_Systems/xlib/*.cpp) Widget_Systems/General/WSdialogs.cpp
 override CXXFLAGS +=  $(shell pkg-config x11 --cflags)
 override CFLAGS += $(shell pkg-config x11 --cflags)
 override LDLIBS += $(shell pkg-config x11 --libs) -lz -lpthread
